requirements.txt
asgiref==3.5.2
Django==4.1.1
django-phonenumber-field==6.4.0
django-recaptcha==3.0.0
django-widget-tweaks==1.4.12
phonenumbers==8.12.54
Pillow==9.2.0
sqlparse==0.4.2
